Make sure gun point is calculated when firing
Gun point calculation is skipped in do_ai_frame if
dist_to_player >= 200 but a bot might still fire a homing weapon
if it is further away.
Skip firing homing weapon without visibility when the gun point is
zero.

This is an original game bug, shown for example in this video:
https://www.youtube.com/watch?v=2dP51znGh7o