Fix LASER_ID constants in D1
The original D1 code didn't have constants for the player laser
weapons. It did have a LASER_ID constant but that one referred to the
robot blue laser. This commit adds the LASER_ID_L1 etc. constants from
D2.
The LASER_ID constant was used in original code that intended to reduce
quad damage, but that never worked since it didn't refer to a player
weapon. The LASER_ID constant and the quad damage reduction code is
taken out in this commit to prevent further confusion.