Improve GL texture preload
Instead of scanning the level for textures, use the paged in status
from paging.c for loading GL textures at the start of a level. This is
more complete than the now removed texture scanning in ogl.c.

Also fixes paging.c in D2 to load the hires textures in hires mode.
And adds reloading the textures when the GL context is destroyed.
And remove the redundant filter_blueship_wing parameter of
ogl_loadbmtexture to simplify the calling code.