Fix recording observer demo
The rear view MFDs fix sets the viewer render type to RT_NONE, but this
breaks the demo system (it needs the viewer orientation but doesn't
write the orientation with render type RT_NONE).

Work around it by temporarily setting the render type to RT_POLYOBJ,
and setting it back to RT_NONE on read.

Fixes: 819190b